rest

+++ src/main/java/com/geek/market/repositories/ProductRepository.java
@@ -5,10 +5,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 
 @Repository
-public interface ProductRepository extends JpaRepository<Product,Integer>, JpaSpecificationExecutor<Product> {
-
-
+public interface ProductRepository extends JpaRepository<Product,Long>, JpaSpecificationExecutor<Product> {
 
 }
